Update How Errors From Server is Handle With Axios
Update how errors from the server are handled with Axios as the error
message is hidden in Axios call, so we will need to check if the error
is Axios error and, if so, choose a different way of accessing it.
However, the test does not know how to handle Axios, so we will need to
keep the original method of getting errors.