Rename Logout to Remove Token on Server
Rename the logout function to Remove Token on the Server, as it will be
used when the user's token expires, and it should remove the token from
the database. So that the code is readable, changing the name will make
it clear to the developer what the function is doing.